fix(services): fix compiler warnings - int to bool conversions and dynamic_cast null check
- PaymentDatabaseUtils: cast aMinutes to qint64 before multiplication
- GUIService: add default case to switch statement on EEventType
- RemoteService:
  - explicitly compare int error to 0 instead of implicit bool conversion
  - add null check after dynamic_cast for terminal settings
  - fix command status to use explicit Error/OK enums

+++ apps/EKiosk/src/DatabaseUtils/PaymentDatabaseUtils.cpp
@@ -1451,9 +1451,10 @@ bool DatabaseUtils::suspendPayment(qint64 aPayment, int aMinutes) {
         return false;
     }
 
-    query->bindValue(
-        ":next_try_date",
-        QDateTime::currentDateTime().addSecs(aMinutes * 60).toString(CIDatabaseProxy::DateFormat));
+    query->bindValue(":next_try_date",
+                     QDateTime::currentDateTime()
+                         .addSecs(qint64(aMinutes) * 60)
+                         .toString(CIDatabaseProxy::DateFormat));
     query->bindValue(":id", aPayment);
 
     if (!query->exec()) {
